Use address clusters to select latest EPC
DLUHC publish EPC data with UPRNs. UPRNs are either submitted by the
assessor or matched to addresess by DLUHC.

Not all certificates have a UPRN. Some UPRNs will be incorrectly
assigned. EPCs also have building reference numbers (BRNs), but it's
relatively easy to find premises with identical addresses but different
BRNs.

For these reasons we deduplicate addresses [0] and assign records an
`address_cluster_id`. If the `address_cluster_id` is the same, then we
think the records describe same premises.

This commit changes the model so that it selects the latest certificate
for a given `address_cluster_id`. If there isn't a cluster, we use the
UPRN, and if there isn't a UPRN we use the BRN.

[0]: https://github.com/centrefornetzero/epc-ppd-address-matching